Move Microsoft.PlatformChannels.Binding.Apple into Platforms to seperate from Microsoft.PlatformChannels.Binding that is Android today
