--- Microsoft.PlatformChannels.Binding.Apple/ApiAdditions.cs
@@ -1,10 +1,0 @@
-//using DotNetPlatformChannels;
-using Foundation;
-using ObjCRuntime;
-//using ObjectiveC;
-
-namespace Microsoft.PlatformChannels.Platform
-{
-	
-	
-}
