--- Microsoft.PlatformChannels.Binding.Apple/Microsoft.PlatformChannels.Binding.Apple.csproj
@@ -1,28 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>net6.0-ios;net6.0-maccatalyst</TargetFrameworks>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <EnableDefaultCompileItems>False</EnableDefaultCompileItems>
-    <RootNamespace>Microsoft.PlatformChannels.Platform</RootNamespace>
-  </PropertyGroup>
-  
-  <PropertyGroup Condition="'$(TargetFramework)' == 'net6.0-ios' OR '$(TargetFramework)' == 'net6.0-maccatalyst' ">
-      <IsBindingProject>True</IsBindingProject>
-      <!--<NoBindingEmbedding>true</NoBindingEmbedding>-->
-	  <!--<SupportedOSPlatformVersion>10.0</SupportedOSPlatformVersion>-->
-  </PropertyGroup>
-
-  <ItemGroup Condition="'$(TargetFramework)' == 'net6.0-ios' OR '$(TargetFramework)' == 'net6.0-maccatalyst' ">
-    <!--<Compile Include="ApiAdditions.cs" />-->
-    <NativeReference Include="..\Platforms\Apple\DotNetPlatformChannels\build\DotNetPlatformChannels.xcframework">
-        <!--<Link>DotNetPlatformChannels.xcframework</Link>-->
-        <Kind>Framework</Kind>
-        <Frameworks></Frameworks>
-        <!--<SmartLink>True</SmartLink>
-        <ForceLoad>True</ForceLoad>
-        <LinkerFlags>-ObjC</LinkerFlags>-->
-    </NativeReference>
-    <ObjcBindingApiDefinition Include="ApiDefinitions.cs" />
-  </ItemGroup>
-</Project>
