+++ Platforms/Apple/Microsoft.PlatformChannels.Binding.Apple/Microsoft.PlatformChannels.Binding.Apple.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFrameworks>net6.0-ios;net6.0-maccatalyst</TargetFrameworks>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <EnableDefaultCompileItems>False</EnableDefaultCompileItems>
+    <RootNamespace>Microsoft.PlatformChannels.Platform</RootNamespace>
+    <IsBindingProject>True</IsBindingProject>
+  </PropertyGroup>
+  
+  <ItemGroup >
+    <NativeReference Include="..\DotNetPlatformChannels\build\DotNetPlatformChannels.xcframework">
+        <Kind>Framework</Kind>
+    </NativeReference>
+    <ObjcBindingApiDefinition Include="ApiDefinitions.cs" />
+  </ItemGroup>
+</Project>
